Cleaned DART waf tools/script

+++ hexapod_dart/src/hexapod.cpp
@@ -95,14 +95,14 @@ namespace robot {
     void Hexapod::_remove_legs()
     {
         std::vector<int> to_remove;
-        for (int i = 1; i < _skeleton->getNumJoints(); i += 3) {
+        for (size_t i = 1; i < _skeleton->getNumJoints(); i += 3) {
             std::vector<int>::iterator it = std::find(_broken_legs.begin(), _broken_legs.end(), i / 3);
             if (it != _broken_legs.end()) {
                 to_remove.push_back(i);
             }
         }
 
-        for (int i = 0; i < to_remove.size(); i++) {
+        for (size_t i = 0; i < to_remove.size(); i++) {
             auto tmp = _skeleton->getJoint(to_remove[i] - i * 3)->getChildBodyNode();
             tmp->removeAllCollisionShapes();
             tmp->removeAllVisualizationShapes();
