+++ hexapod_dart/src/hexapod_dart_simu.cpp
@@ -303,11 +303,11 @@ void HexapodDARTSimu::_check_duty_cycle()
 {
     auto rob = this->robot();
 
-    for (unsigned i = 0; i < 6; ++i) {
+    for (size_t i = 0; i < 6; ++i) {
         std::string leg_name = "leg_" + std::to_string(i) + "_3";
         dart::dynamics::BodyNodePtr body_to_check;
         // TO-DO: Maybe there's a cleaner way to get the body
-        for (int j = 0; j < rob->skeleton()->getNumBodyNodes(); j++) {
+        for (size_t j = 0; j < rob->skeleton()->getNumBodyNodes(); j++) {
             auto bd = rob->skeleton()->getBodyNode(j);
             if (leg_name == bd->getName())
                 body_to_check = bd;
